requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
aioredis==1.3.1
asgiref==3.5.0; python_version >= '3.7'
async-timeout==4.0.2; python_version >= '3.6'
attrs==21.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
autobahn==22.2.2; python_version >= '3.7'
automat==20.2.0
backports.zoneinfo==0.2.1; python_version < '3.9'
certifi==2021.10.8
cffi==1.15.0
channels-redis==3.4.0
channels==3.0.4
charset-normalizer==2.0.12; python_version >= '3'
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==36.0.2
daphne==3.0.2
django-cors-headers==3.11.0
django-filter==21.1
django-storages==1.12.3
django==4.0.3
djangorestframework==3.13.1
drf-jwt==1.19.2
drf-nested-routers==0.93.4
drf-yasg==1.20.0
hiredis==2.0.0; python_version >= '3.6'
hyperlink==21.0.0
idna==3.3; python_version >= '3'
importlib-metadata==4.11.3; python_version < '3.10'
incremental==21.3.0
inflection==0.5.1; python_version >= '3.5'
itypes==1.2.0
jinja2==3.0.3; python_version >= '3.6'
markdown==3.3.6
markupsafe==2.1.1; python_version >= '3.7'
msgpack==1.0.3
packaging==21.3; python_version >= '3.6'
psycopg2==2.9.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
pycparser==2.21
pyjwt==2.3.0
pyopenssl==22.0.0
pyparsing==3.0.7; python_version >= '3.6'
python-dotenv==0.19.2
pytz==2022.1
requests==2.27.1
ruamel.yaml.clib==0.2.6; python_version < '3.11' and platform_python_implementation == 'CPython'
ruamel.yaml==0.17.21; python_version >= '3'
service-identity==21.1.0
setuptools==60.10.0; python_version >= '3.7'
six==1.16.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
sqlparse==0.4.2; python_version >= '3.5'
twisted[tls]==22.2.0; python_full_version >= '3.6.7'
txaio==22.2.1; python_version >= '3.6'
typing-extensions==4.1.1; python_version >= '3.6'
uritemplate==4.1.1; python_version >= '3.6'
urllib3==1.26.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
zipp==3.7.0; python_version >= '3.7'
zope.interface==5.4.0; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4'
